Update 5 Platformer_Practice
Added a main menu with buttons to start a game, continue a game, edit game options, and quit. The main menu leads into the game through the first button. Will need to make sure the game screen clears before returning from the game over menu to the main menu. Will need to work on this later.