Update 6 Platformer_Practice
Added an in-game menu that can be opened and closed by pressing the Escape key. Want to add a semi-transparent surface to contain the buttons. Right now it allows the user to pause their game (by opening the menu), continue it, access options (not currently but the button is there), and exit the game.