Update 7 Platformer_Practice
Added two powerup items to be picked up by the player. The first is the "boost" that launches a player upwards at a high speed. The second is the "feather" powerup that allows the player to jump a little higher and fall much more slowly. The feather has a limited number of uses and the boost happens once and immediately. Will probably need to clean up code, and want to add powerups that are time based.