Update 9 Platformer_Practice
Changed the "Coins" group to the "Items" group. Added a save point and created a way for the player to interact with the save point by pressing "e" on the keyboard. Right now this does nothing but print a statement, but will eventually allow for saving the game as well as interacting with other items. Will need to add a save function, perhaps a text document that catalogues the player's progress and that can be accessed later to load the game.