Update 10 Platformer_Practice
Made some progress on the saving function of the game. When the player interacts with the Save Point, it will print out a list of all variable values for all objects (player, platforms, items) in the game. Also changed around some things in classes and added "object types" that allows for more easy tracking of objects and for objects to more easily interact with one another.