Update 11 Platformer_Practice
Made some progress on the save function and the load function. The load function currently can open and read a file. The save function right now can retrieve the variables of an object, and the values of those variables. This should eventually allow for the writing of those two things to a file by the save function, which can then be opened and read later by the load function and hopefully used to regenerate the objects of the game.