Update 12 Platformer_Practice
Worked more on the Save and Load functions. The game will save information about objects that allows their object type and other values to be identified. The Load function can decipher this recorded data and currently returns what type of object is in each line of the save file. Eventually this will allow for the iterative loading of objects where the object type is identified, the object is placed into the correct sprite group, and where the parameters are then used to actually generate the object.