Update 13 Platformer_Practice
The load function can now add objects back into their groups from the save file, and it can now reproduce Savepoints on the screen where they were originally when the player saved the game. Will need to do the same for the Player, platforms, and all other items now. Will need to clean up the save and load functions. Happy with the progress made so far.