Update 14 Platformer_Practice
Added a lot of different color values. Changed some classes to make it easier to create objects from them when loading the game from a save file. Platforms can be loaded and will resume moving as they did when the game was saved. Made some other progress in the Load function (like loading the saved score) and did some work on the Powerups class. Need to work on loading powerups.