Update 15 Platformer_Practice
Added a life mechanic to the game. The character begins the game with an extra life, which is consumed if they fall off the screen. They will then be catapulted back up into the gameplay area. If they run out of lives by falling off the screen too many times, they lose the game. More lives can be collected, and they are generated as one of the Powerups.