Update 18 Platformer_Practice
Fleshed out the Options menu a little bit more with buttons that represent Sound, Graphics, and Controls settings. Also added a fancy border to the Main Menu and the Options Menu. Another thing to work on is adding the ability for the Player to actually change the game settings. Currently thinking about adding an NPC who sells Upgrades to the Player as well.