Update 19 Platformer_Practice
Made some crude progress on the Sound, Graphics, and Controls menus. Will need to simplify the border creation for all menus, right now it relies on defining four rectangles and blitting them to the screen individually. Could probably solve this with a function, or make the border a sprite. Will see how things progress. For now the Sound, Graphics, and Controls menus are not implemented and have no use to the Player.