Update 20 Platformer_Practice
Simply fleshed out the Sound, Graphics, and Controls menus a little bit more. Clicking the "Options" button in the Main Menu brings you to a selection of these menus, and clicking each of these menus now brings you to a new individual menu. Each will in turn have their own settings that can be changed so the Player can customize the game more to their liking.