Update 21 Platformer_Practice
Tentatively added a Merchant NPC class. Right now it needs a lot of work, and the only thing it can do is automatically choose to color itself one of three ways and then fall off the screen. Hoping to make it more like the Player, as well as add salesman functionality.